requirements/development.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
SQLModel==0.0.16
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
black==23.11.0
isort==5.12.0
flake8==6.1.0 
requests==2.25.1
dependency-injector==4.39.1
asyncpg==0.25.0
passlib==1.7.4